requirements.txt
Django==4.2.7
feedparser==6.0.10
requests==2.31.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.7
elasticsearch==8.11.0
slack-sdk==3.26.1
python-telegram-bot==20.7
django-celery-beat==2.5.0
django-celery-results==2.5.1 